requirements.txt
asgiref==3.11.0
attrs==25.4.0
chardet==5.2.0
Django==5.2.8
django-debug-toolbar==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
flake8==7.3.0
flake8-quotes==3.4.0
flake8-variables-names==0.0.6
gitdb==4.0.12
GitPython==3.1.45
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
pep8-naming==0.15.1
pillow==12.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.37.0
rich==14.2.0
rpds-py==0.29.0
setuptools==80.9.0
smmap==5.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
psycopg==3.3.1
psycopg-binary==3.3.1
dotenv
